sof: math: move iir_df2t function to src/math
Moved iir_df2t() from src/audio/eq_iir to src/math as a common library for
IIR and Crossover usage.
