+++ src/audio/eq_iir/CMakeLists.txt
@@ -1,3 +1,3 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
-add_local_sources(sof eq_iir.c iir.c iir_generic.c iir_hifi3.c)
+add_local_sources(sof eq_iir.c iir.c)
